feat: add outlier ploting to outs
